Personalapp/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "google-genai>=1.73.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
